# docs/openapi.json
{
  "components": {
...
    "responses": {
      "BadRequest": {
        "description": "Bad Request",
        "examples": [
          {
            "error": "Bad Request",
            "message": "Invalid input data.",
            "success": false
          }
        ],
        "properties": {
          "error": {
            "type": "string"
          },
          "message": {
            "type": "string"
          },
          "success": {
            "type": "boolean"
          }
        },
        "required": [
          "success",
          "message",
          "error"
        ],
        "type": "object"
      },
      "Conflict": {
        "description": "Conflict",
        "examples": [
          {
            "error": "Conflict.",
            "message": "The resource already exists.",
            "success": false
          }
        ],
        "properties": {
          "error": {
            "description": "Error details, if any",
            "example": "Conflict.",
            "type": "string"
          },
          "message": {
            "description": "A message providing additional information",
            "example": "The resource already exists.",
            "type": "string"
          },
          "success": {
            "description": "Indicates the request failed due to a conflict in the request",
            "example": false,
            "type": "boolean"
          }
        },
        "type": "object"
      },
      "Forbidden": {
        "description": "Forbidden",
        "examples": [
          {
            "error": "Forbidden.",
            "message": "You do not have permission to access this resource.",
            "success": false
          }
        ],
        "properties": {
          "error": {
            "description": "Error details, if any",
            "example": "Forbidden.",
            "type": "string"
          },
          "message": {
            "description": "A message providing additional information",
            "example": "You do not have permission to access this resource.",
            "type": "string"
          },
          "success": {
            "description": "Indicates the request failed due to insufficient permissions",
            "example": false,
            "type": "boolean"
          }
        },
        "type": "object"
      },
...
      "NotFound": {
        "description": "Not Found",
        "examples": [
          {
            "error": "Not Found.",
            "message": "Resource not found.",
            "success": false
          }
        ],
        "properties": {
          "error": {
            "description": "Error details, if any",
            "example": "Not Found.",
            "type": "string"
          },
          "message": {
            "description": "A message providing additional information",
            "example": "Resource not found.",
            "type": "string"
          },
          "success": {
            "description": "Indicates the requested resource could not be found",
            "example": false,
            "type": "boolean"
          }
        },
        "type": "object"
      },
      "Unauthorized": {
        "description": "Unauthorized",
        "examples": [
          {
            "error": "Unauthorized.",
            "message": "Unauthorized access.",
            "success": false
          }
        ],
        "properties": {
          "error": {
            "description": "Error details, if any",
            "example": "Unauthorized.",
            "type": "string"
          },
          "message": {
            "description": "A message providing additional information",
            "example": "Unauthorized access.",
            "type": "string"
          },
          "success": {
            "description": "Indicates the request failed due to authentication failure",
            "example": false,
            "type": "boolean"
          }
        },
        "type": "object"
      },
...
    },
    "schemas": {
...
      "ProjectStateRequest": {
        "additionalProperties": false,
        "description": "Represents a request to create or update a project state.",
        "examples": [
          {
            "name": "In Progress"
          },
          {
            "name": "Completed"
          },
          {
            "name": "On Hold"
          }
        ],
        "properties": {
          "name": {
            "description": "Name of the project state.",
            "example": "In Progress",
            "type": "string"
          }
        },
        "required": [
          "name"
        ],
        "title": "ProjectStateRequest",
        "type": "object"
      },
      "ProjectStateResponse": {
        "additionalProperties": false,
        "description": "Represents the response structure for a project state.",
        "examples": [
          {
            "created_at": "2025-02-07T12:30:45.123Z",
            "created_by": "admin@example.com",
            "id": "3f57b60c-5d4f-4b0b-bf10-2c9b9e2c7d4a",
            "name": "In Progress",
            "project_id": "ae12d34b-567c-89ef-01gh-234567890ijk",
            "sequence": 2,
            "updated_at": "2025-02-07T15:45:30.456Z"
          },
          {
            "created_at": "2025-02-05T08:15:00.789Z",
            "created_by": "manager@example.com",
            "id": "fa68c12d-34ef-4a78-91bc-5e2a67890abc",
            "name": "Completed",
            "project_id": "ae12d34b-567c-89ef-01gh-234567890ijk",
            "sequence": 3,
            "updated_at": "2025-02-06T10:00:25.456Z"
          }
        ],
        "properties": {
          "created_at": {
            "description": "Timestamp indicating when the state was created.",
            "example": "2025-02-07T12:30:45.123Z",
            "format": "date-time",
            "type": "string"
          },
          "created_by": {
            "description": "Email of the user who created the state.",
            "example": "admin@example.com",
            "format": "email",
            "type": "string"
          },
          "id": {
            "description": "Unique identifier for the project state.",
            "example": "3f57b60c-5d4f-4b0b-bf10-2c9b9e2c7d4a",
            "format": "uuid",
            "type": "string"
          },
          "name": {
            "description": "Name of the project state.",
            "example": "In Progress",
            "type": "string"
          },
          "project_id": {
            "description": "Unique identifier of the associated project.",
            "example": "ae12d34b-567c-89ef-01gh-234567890ijk",
            "format": "uuid",
            "type": "string"
          },
          "sequence": {
            "description": "Order or sequence number of the state.",
            "example": 2,
            "format": "int32",
            "type": "integer"
          },
          "updated_at": {
            "description": "Timestamp indicating when the state was last updated.",
            "example": "2025-02-07T15:45:30.456Z",
            "format": "date-time",
            "type": "string"
          }
        },
        "required": [
          "id",
          "name",
          "created_by",
          "project_id",
          "sequence",
          "created_at",
          "updated_at"
        ],
        "title": "Project State Response",
        "type": "object"
      },
...
    },
    "securitySchemes": {
      "BearerAuth": {
        "bearerFormat": "JWT",
        "description": "JWT Token",
        "scheme": "bearer",
        "type": "http"
      }
    }
  },
  "info": {
    "contact": {
      "email": "support@projectmanagement.com",
      "name": "Project Management API Team",
      "url": "http://projectmanagement.com"
    },
    "description": "API for managing clients, projects, and issues to demonstrate the features of OpenAPI 3.1 specification.",
    "license": {
      "name": "Apache 2.0",
      "url": "https://www.apache.org/licenses/LICENSE-2.0.html"
    },
    "termsOfService": "http://example.com/terms/",
    "title": "Project Management API",
    "version": "1.0.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/project/{project_id}/state": {
      "post": {
        "description": "Create a new state for a specific project.",
        "operationId": "createProjectState",
        "parameters": [
          {
            "description": "Unique identifier of the project.",
            "in": "path",
            "name": "project_id",
            "required": true,
            "schema": {
              "example": "123e4567-e89b-12d3-a456-426614174000",
              "format": "uuid",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ProjectStateRequest"
              }
            }
          },
          "description": "State data to create a new project state.",
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "properties": {
                    "data": {
                      "$ref": "#/components/schemas/ProjectStateResponse"
                    },
                    "message": {
                      "example": "Project state created successfully.",
                      "type": "string"
                    },
                    "success": {
                      "example": true,
                      "type": "boolean"
                    }
                  },
                  "type": "object"
                }
              }
            },
            "description": "Successfully created project state."
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/responses/BadRequest"
                }
              }
            },
            "description": "Bad request due to invalid input."
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/responses/Unauthorized"
                }
              }
            },
            "description": "Unauthorized request."
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/responses/Forbidden"
                }
              }
            },
            "description": "Forbidden request."
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/responses/NotFound"
                }
              }
            },
            "description": "Project not found."
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/responses/Conflict"
                }
              }
            },
            "description": "Conflict due to an existing project state with the same identifier."
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/responses/UnprocessableEntity"
                }
              }
            },
            "description": "Unprocessable entity due to validation errors."
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/responses/InternalServerError"
                }
              }
            },
            "description": "Internal server error."
          }
        },
        "summary": "Create a new project state",
        "tags": [
          "ProjectState"
        ]
      },
      "put": {
        "description": "Update the sequence of states by providing an ordered list of state UUIDs.",
        "operationId": "updateStateSequence",
        "parameters": [
          {
            "description": "Unique identifier of the project.",
            "in": "path",
            "name": "project_id",
            "required": true,
            "schema": {
              "example": "123e4567-e89b-12d3-a456-426614174000",
              "format": "uuid",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateStatesSequenceRequest"
              }
            }
          },
          "description": "The request body containing the updated sequence of project states, ensuring they are reordered according to the specified structure.",
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "properties": {
                    "message": {
                      "example": "State sequence updated successfully.",
                      "type": "string"
                    },
                    "success": {
                      "example": true,
                      "type": "boolean"
                    }
                  },
                  "type": "object"
                }
              }
            },
            "description": "Successfully updated the state sequence."
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/responses/BadRequest"
                }
              }
            },
            "description": "Bad request due to invalid input."
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/responses/Unauthorized"
                }
              }
            },
            "description": "Unauthorized request."
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/responses/Forbidden"
                }
              }
            },
            "description": "Forbidden request."
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/responses/NotFound"
                }
              }
            },
            "description": "One or more states not found."
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/responses/InternalServerError"
                }
              }
            },
            "description": "Internal server error."
          }
        },
        "summary": "Update State Sequence",
        "tags": [
          "ProjectState"
        ]
      }
    },
...
    "/project/{project_id}/states": {
      "get": {
        "description": "Retrieve a paginated list of states for a specific project.",
        "operationId": "listProjectStates",
        "parameters": [
          {
            "description": "Unique identifier of the project.",
            "in": "path",
            "name": "project_id",
            "required": true,
            "schema": {
              "example": "123e4567-e89b-12d3-a456-426614174000",
              "format": "uuid",
              "type": "string"
            }
          },
          {
            "description": "The page number for pagination.",
            "in": "query",
            "name": "page",
            "required": false,
            "schema": {
              "default": 1,
              "type": "integer"
            }
          },
          {
            "description": "The number of items per page.",
            "in": "query",
            "name": "page_size",
            "required": false,
            "schema": {
              "default": 10,
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "properties": {
                    "data": {
                      "items": {
                        "$ref": "#/components/schemas/ProjectStateResponse"
                      },
                      "type": "array"
                    },
                    "message": {
                      "example": "Project states retrieved successfully.",
                      "type": "string"
                    },
                    "meta": {
                      "$ref": "#/components/schemas/Meta"
                    },
                    "success": {
                      "example": true,
                      "type": "boolean"
                    }
                  },
                  "type": "object"
                }
              }
            },
            "description": "Successfully retrieved project states."
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/responses/BadRequest"
                }
              }
            },
            "description": "Bad request due to invalid input."
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/responses/Unauthorized"
                }
              }
            },
            "description": "Unauthorized request."
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/responses/Forbidden"
                }
              }
            },
            "description": "Forbidden request."
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/responses/NotFound"
                }
              }
            },
            "description": "Project or states not found."
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/responses/InternalServerError"
                }
              }
            },
            "description": "Internal server error."
          }
        },
        "summary": "List all project states with pagination",
        "tags": [
          "ProjectState"
        ]
      }
    },
...
  },
  "security": [
    {
      "BearerAuth": []
    }
  ],
  "servers": [
    {
      "description": "Test server",
      "url": "https://test.projectmanagement.com/api/v1"
    },
    {
      "description": "Development server",
      "url": "https://dev.projectmanagement.com/api/v1"
    },
    {
      "description": "Production server",
      "url": "https://api.projectmanagement.com/api/v1"
    }
  ],
...
}